2nd_Assigment/2d_plot_diffusion_todo/requirements.txt
# Updated requirements.txt
scikit-learn>=1.6.0  # Supports Python 3.13 [2][5]
numpy>=2.1.0  # Python 3.13 compatibility [3][6]
ipython>=8.12.0
jupyter>=1.0.0
matplotlib>=3.8.0  # Modern version for Python 3.13
torch-ema>=0.3
tqdm>=4.64.1
jupyterlab>=4.0.2
jaxtyping>=0.2.20
pytorch_lightning>=2.2.0
dotmap>=1.0.0
